requirements.txt
aiogram==3.7.0
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
loguru==0.7.2 